+++ CHANGELOG.md
@@ -7,3 +7,5 @@ CHANGELOG
 - Blocks have a widget instead of widgets having blocks #22
 - Introduced `Sprite` shape
 - Added `TextShape` shape which renders fonts.
+- Added `ImageShape` widget to render images on the canvas #36.
+